requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==5.5.0
async-timeout==5.0.1
attrs==24.2.0
base58==2.1.1
bitarray==2.9.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cytoolz==0.12.3
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.7.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
hexbytes==0.3.1
idna==3.10
importlib-metadata==6.11.0
ipfshttpclient==0.8.0a2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lru-dict==1.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multiaddr==0.0.9
multidict==6.1.0
narwhals==1.18.3
netaddr==1.3.0
numpy==1.26.4
packaging==23.2
pandas==2.2.3
parsimonious==0.8.1
Pillow==9.5.0
propcache==0.2.1
protobuf>=3.20,<6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
referencing==0.35.1
requests==2.31.0
rich==13.9.4
rlp==2.0.1
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
streamlit==1.25.0
tenacity==8.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
urllib3==2.2.3
validators==0.34.0
varint==1.0.2
web3==5.30.0
websockets==9.1
yarl==1.18.3
zipp==3.21.0
